requirements.txt
Flask==2.0.1
Flask-MySQLdb==2.0.0
Flask-Login==0.5.0
python-dotenv==0.19.0
Pillow==8.4.0
werkzeug==2.0.1